* [feat] increase the precision from float to double * [bugfix] print the warning/errors in stderr * [bugfix] perform addition checks (slower) for some files which generated segfault

+++ MST-Pathfinder/extern.h
@@ -14,16 +14,16 @@
 typedef struct _SEDGE
 {
   int i, j;
-  float v;
+  double v;
 } SEDGE;
 
 
 /* VARIABLES */
 extern int q;		//parametro q para indicar el número de arcos max que puede tener un camino
-extern float **pfnet;	//si pfnet[i][j]=TRUE hay un enlace entre los nodos i y j
+extern double **pfnet;	//si pfnet[i][j]=TRUE hay un enlace entre los nodos i y j
 extern int num_nodos;	//numero de nodos de la matriz
 extern int num_edges;	// number of edges in the matrix
-extern float **pesos_originales;
+extern double **pesos_originales;
 extern SEDGE **vedges;
 extern int *compConnexe;
 extern int *compConnexeRank;
@@ -32,8 +32,8 @@ extern int *sameValuesNodes;
 
 
 /* FUNCIONES Y PROCEDIMIENTOS*/
-extern float ** Matriz_dinamica_float(int nfil, int ncol);	//reserva memoria para una matriz de flotantes
-extern void Libera_matriz_dinamica_float (float **matriz, int nfils);	//libera memoria de una matriz de flotantes
+extern double ** Matriz_dinamica_float(int nfil, int ncol);	//reserva memoria para una matriz de flotantes
+extern void Libera_matriz_dinamica_float (double **matriz, int nfils);	//libera memoria de una matriz de flotantes
 extern void Input(int argc, char *argv[]);	//coge los datos de entrada y los guarda en las respectivas variables
 extern void Programacion_Dinamica ();	//Función que resuelve el problema mediante programación dinamica
 static int compedges_min(const void *_e1, const void *_e2);
